FSM Refactor: Extract out sliding and walking states into separate files
