+++ Assets/PenguinQuest/Code/Controllers/Fsm/FsmState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8bd76be2270b4848aa57c6f59e5c0b4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
