+++ Assets/PenguinQuest/Code/Controllers/Fsm/PenguinSlidingState.cs
@@ -1,0 +1,37 @@
+﻿using UnityEngine;
+using PenguinQuest.Data;
+using PenguinQuest.Controllers.Fsm;
+
+
+namespace PenguinQuest.Controllers.Fsm
+{
+    public class SlidingState : FsmState
+    {
+        public SlidingState(string name) : base(name) { }
+
+        public override void Enter()
+        {
+
+        }
+
+        public override void Exit()
+        {
+
+        }
+
+        public override void FixedUpdate()
+        {
+
+        }
+
+        public override void LateUpdate()
+        {
+
+        }
+
+        public override void Update()
+        {
+
+        }
+    }
+}
