+++ Assets/PenguinQuest/Code/Controllers/Fsm/PenguinSlidingState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d179bf6dd412682459ea15e0b48f95b4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
