+++ Assets/PenguinQuest/Code/Controllers/Fsm/PenguinWalkingState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a729b70da6fe6ca44a4aa69513f05b97
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
